Rewrite runner and all clients to use IPC.
This major commit refactors the benchmark runner and client drivers to
use async IPC.

This change enables users to implement client drivers in their language
of choice.

To do so, they must implement deserialization and serialization of our
FlatBuffers types, as exemplified by the rewritten Golang client
drivers for Cassandra, Mongo, Redis, and the ephemeral key/value
datastore.

Note that the FlatBuffers schema was rewritten to better support
this new architecture.

Additionally, developer UX has improved as a result of this rewrite.